requirements.txt
tensorflow>=2.15.0
numpy>=1.26.0,<2.0
opencv-python>=4.8.0
xgboost>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
Pillow>=10.0.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0